reorg views
