+++ nbcli/views/__init__.py
@@ -1,12 +1,24 @@
 """Load views used by nbprint."""
 
 from nbcli.views.tools import BaseView                          # noqa: F401
-from nbcli.views.device import DcimDevicesView                  # noqa: F401
-from nbcli.views.interfaces import DcimInterfacesView           # noqa: F401
-from nbcli.views.racks import DcimRacksView, DcimRUsView        # noqa: F401
-from nbcli.views.locations import DcimLocationsView             # noqa: F401
-from nbcli.views.sites import DcimSitesView                     # noqa: F401
-from nbcli.views.aggregates import IpamAggregatesView           # noqa: F401
-from nbcli.views.ip_addresses import IpamIpAddressesView        # noqa: F401
-from nbcli.views.ip_prefixes import IpamPrefixesView            # noqa: F401
-from nbcli.views.object_changes import ExtrasObjectChangesView  # noqa: F401
+
+from nbcli.views.circuits import CircuitsProvidersView          # noqa: F401
+from nbcli.views.circuits import CircuitsCircuitTypesView       # noqa: F401
+from nbcli.views.circuits import CircuitsCircuitsView           # noqa: F401
+
+from nbcli.views.dcim import DcimDevicesView                  	# noqa: F401
+from nbcli.views.dcim import DcimInterfacesView           		# noqa: F401
+from nbcli.views.dcim import DcimRacksView        				# noqa: F401
+from nbcli.views.dcim import DcimRUsView        				# noqa: F401
+from nbcli.views.dcim import DcimLocationsView             		# noqa: F401
+from nbcli.views.dcim import DcimSitesView                     	# noqa: F401
+
+from nbcli.views.extras import ExtrasConfigContextsView  		# noqa: F401
+from nbcli.views.extras import ExtrasObjectChangesView  		# noqa: F401
+
+from nbcli.views.ipam import IpamAggregatesView           		# noqa: F401
+from nbcli.views.ipam import IpamIpAddressesView        		# noqa: F401
+from nbcli.views.ipam import IpamPrefixesView            		# noqa: F401
+
+from nbcli.views.tenancy import TenancyTenantGroupsView         # noqa: F401
+from nbcli.views.tenancy import TenancyTenantsView              # noqa: F401
